This patch enables compilation of the example benchmark for gem5 run.
Previous CMake compilation system in this repo does not copmile any
benchmarks for gem5 use.
This patch removes unuse options, such as GEM5_PATH, and adds a
syscall_wrapper library to the repo to trigger linux system emulation
in gem5 SE mode for syscalls.
This patch also fixes the previous gem5 m5op assembly because the
previous one was triggering event queue unsync problem during the
test.
This patch also removes the linking to the boot.ld file in startup.